qa: Honor cephfs config from yaml across test cases
The cephfs_test_runner wasn't honoring the cephfs
config passed via yaml as below.

------
ceph:
  cephfs:
    allow_referent_inodes: true
    max_mds: 2
    fs:
      - name: a
        max_mds: 3
      - name: b
------

For each test case the filesystem is newly created and destroyed as
part of setup and teardown. The filesystem creation is taken care
by the class 'Filesystem' in which the 'fs_config' isn't initialised
from yaml. Hence the filesystem is being created with default configs
without honoring the configs passed from yaml. This patch fixes the
same.

Fixes: https://tracker.ceph.com/issues/71649
Signed-off-by: Kotresh HR <khiremat@redhat.com>