osd/scrub: allow longer waits for replicas to respond
Increase the two timeouts associated with replica
responses to scrub requests.

This solves the issue of, when a cluster event causes
some form of repeering (e.g. an OSD in the active set
is down), having a request time out before the new
interval is established. This scenario does not
lead to any real data loss or crashes, but it does
result in log warnings (and failed tests).

Fixes: https://tracker.ceph.com/issues/68698

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>