os/bluestore: Writer; add write_no_read
Added new function to the Writer toolkit that performs
modification of object data without reading existing data.
It is necessary to fix cases of BlueStore crashing on write
when underlying data is corrupted. We do not allow failures
from ObjectStore::queue_transactions().
The solution here is to leave previous data untouched and
defer handling -EIO to ObjectStore::read().

Signed-off-by: Adam Kupczyk <akupczyk@ibm.com>