rgw: introduce decrypt-mode to allow sse-s3 decryption in multisite
Passing decrypt-mode rather than a bool skip-decrypt would allow the
dest zone to pass "skip-except-sse-s3" when we have the src and dest
bucket tre not he same and passing "skip" when they are.
so with that, source zone can make the right call based on the string
provided by the decrypt-mode param.

Signed-off-by: Seena Fallah <seenafallah@gmail.com>