mgr/cephadm: clear CEPHADM_DAEMON_PREPARE_CREATE_FAIL warnings
This is a new warning with the parallel deployment work
(it would have fallen under the general CEPHADM_APPLY_SPEC_FAIL
before) so we need to make sure to clear it along with the
other deployment related health warnings. Otherwise, once it
happens once the warning sticks around forever even if
the issue has been resolved.

Signed-off-by: Adam King <adking@redhat.com>