mgr/cephadm: skip keepalive deployment until haproxy has been deployed
We need the haproxy daemons in order to be able to populate
the keepalive conf with targets for keepalive to check
the health of. Before deploying in parallel, this worked
as we would do prepare_create and deployment for each
daemon sequentially and would deploy haproxy before keepalived.
Now that we're doing all the prepare_create steps first
and then deploying in parallel, keepalived's prepare_create
will fail unless we've already  deployed haproxy daemons
for that service.

Signed-off-by: Adam King <adking@redhat.com>