mgr/cephadm: introduce deployment ordering for services
Some services don't work well with deploying every
daemon in parallel and need deployment to follow
a certain ordering. This commit sets up the framework
for allowing services to specify the order in which
its daemons should be deployed. The expectation is
that the function will return a dict mapping integers
to a list of daemons. When we deploy, we first deploy
everything in the "0" entry, then "1" etc. The integers
themselves are fairly arbitrary, we just need something
sortable.

Signed-off-by: Adam King <adking@redhat.com>