test/rgw/notifications: prevent client retries to avoid duplicates
if the RGW is slow, and the client retry, it may cause test to fail
since the number of notifications would be off.
in addition, in slow RGW, we need to verify that the expiry time did
not pass before checking the queue, so we see the expected number of
entries in the queue before they expire.

Fixes: https://tracker.ceph.com/issues/70756

Signed-off-by: Yuval Lifshitz <ylifshit@ibm.com>
(cherry picked from commit dd6d4f0bbc9f0dcbb09ee0230e2f56d7c0e05439)