qa: get_rand_pg_acting_set needs to wait for pool to create PGs
There is a race hazard in the OSD thrasher which tests if a pool
exists and then queries the PGs in the pool. It is possible that
a pool exists (has been added to OSDMap) but the PGs have not
been created yet (by the OSDs). Add a sleep/retry to mitigate
the race.

Fixes: https://tracker.ceph.com/issues/70818

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>
(cherry picked from commit cbaadf521919e28dc2d916088a106304a18830f9)