crimson/common/smp_helpers: cleanup sharded_map_seq
Let PGShardManager::invoke_on_each_shard_seq pass the local shard_services
instance instead of using an additional helper.

The downside of dropping the generic sharded_map_seq helper is that it is
able to support *any* (seastar::)sharded object. However, as shard_services
is the only user of it - directly using the local instance without the
helper seems easier to read.

Signed-off-by: Matan Breizman <mbreizma@redhat.com>