doc/rados: clarify "upmap_max_deviation"
Clarify the threshold set by "upmap_max_deviation" and add the
information about this configurable that is currently in
src/pybind/mgr/balancer/module.py to src/common/options/global.yaml.in,
so that it will be accessible by means of ".. confval::" declarations.

In the backport of global.yaml.in to the Reef release branch, there was
a merge conflict. I removed all of the configurables from the merge
conflict except the one explicitly referred to here
(upmap_max_deviation). This could mean that there are other
configurables in the main branch that should be backported to the Reef
release branch, but that is a matter for another PR.

Fixes: https://tracker.ceph.com/issues/71796

Signed-off-by: Zac Dover <zac.dover@proton.me>
(cherry picked from commit 80de0a3c92d54d644a0ea3d8120f1f8f445ea2e0)