msg/async/ProtocolV2: Speed up connection logic when peer restarts
When a client is restarted, it loses its state including global_seq and
gets a new client_cookie. This creates an issue during reconnection
because the server has an existing connection with a higher
global_seq value, causing it to reject the new connection as "stale"
with the error:

  "this is a stale connection, peer_global_seq="

This commit adds detection logic in
ProtocolV2::handle_existing_connection() to identify client restarts
by checking:

  - exproto->peer_global_seq > peer_global_seq
    (global_seq should only increase during a session; a decrease
    strongly suggests a restart)

  - client_cookie has changed (client generated a new cookie)

  - existing connection state is not Ready or Standby

When all three conditions are met, the server drops the existing connection
and accepts the new one. This leads to faster reconnection, especially
in cases like Monitor restarts, helping clients rejoin the quorum
without delay and avoiding spurious MON_NETSPLIT warnings.

It also prevents reliance on server-side reconnect attempts or waiting
for the client's global_seq to catch up.

Note: without this patch, it'll take longer
for the reconnection to happen, eventually the server will
also act as a client and send_reconnect since it has server_cookie.
However, this could be 15 seconds later, since, the server probably
was hit with a _fault max_back_off when it was trying to connect
to the client when the client was still down and has not yet
restart. This patch basically does not wait for the reconnection
to happen and intelligently make decisions in dropping existing
connection or not based on 3 conditions.

Fixes: https://tracker.ceph.com/issues/71344

Signed-off-by: Kamoltat Sirivadhna <ksirivad@redhat.com>
(cherry picked from commit 1bd8e0c7ab282a081d71c8e49f72dd6bd4551f2b)