doc/radosgw: line edit bucket_logging.rst
Edit doc/radosgw/bucket_logging.rst so that it is not solecistic and so
that its punctuation is corrected and its use of articles is corrected.
This file remains in my judgment demotic and maybe demotic enough to
warrant another editorial pass in the future.

Fixes: https://tracker.ceph.com/issues/70758

Signed-off-by: Zac Dover <zac.dover@proton.me>
(cherry picked from commit c01eb372326be49cbb76d602eabba951663979ec)