os/bluestore: eliminate metadata duplication in OSD dir and bdev label
This change eliminates the duplication of OSD metadata written to both
the bdev label and the OSD directory.

Key changes:
1. BlueStore::write_meta() now avoids writing metadata to the OSD dir
   if bdev labels are supported.
2. For devices that do not support bdev labels, we now automatically
   migrate metadata into the label during mkfs.

Note:
In practice, all devices used with BlueStore can support labeling by
writing to the first 4KB of the device. The earlier assumption that
NVMe or PMEM devices might not support this turned out to be
unnecessary. This cleanup prepares for removing support_bdev_label()
checks entirely in a future change.

Signed-off-by: Jaya Prakash <jayaprakash@ibm.com>