rgw: handle versioned object without current instance
A few fixes and changes to deal with object that has only instances
in snapshots.
The olh in a case where it doesn't have any instance to point at will
now point at an instance with an empty instance, and have the
delete_marker field set to true.
We also now keep track of the latest instance in snapshot that was
deleted in a later snapshot. We call it the "first" because it is
first in the listing order. It is used to track existence of hidden
instances. When the first_hidden instane is deleted we search
for a previous instance to set as the new first_hidden. If one is
not found (and the olh is set to exists=false) then the olh can
be removed.
We also add a new op to the olh log, which is basically a no-op. It
is needed to complete an instance removal with some operation so that
the 'pending' attr that was created on the olh is removed. Even if we
don't remove the instance itself (because the instance is part of
an existing snapshot), we need to complete the operation and delete
the attribute.

Signed-off-by: Yehuda Sadeh <ysadehwe@ibm.com>