script: add redmine-upkeep script
This script intends to do all Redmine upkeep related tasks.

It's organized so that a series of transformations are run on any issue under
examination. These include things like updating the "Merge Commit" and "Fixed
In" fields when the associated PR is merged. The "Released In" field is also
set when the associated merge commit is part of some non-RC release.

Transformations can be run in one of three ways:

* An explicit --issue switch specifies a comma separated list of issues to
  operate on.

* An explicit --revision-range switch specifies a Git revision range to lookup
  first-parent merge commits. For each of those merge commits, try to find
  the associated issue with "Merge Commit" equal to that merge commit. If found,
  operate on that issue. This is intended to allow updating the "Released In"
  field whenever a tag is pushd to ceph.git.

Future work:

* Incorporate backport-create-issue into this script.

* Also update "Target version" if straightforward. This is protected by redmine
  whenever a version is locked/closed.

* Anything else we desire to automate!

Signed-off-by: Patrick Donnelly <pdonnell@ibm.com>