mon/MgrStatMonitor.cc: cleanup handle_conf_change
Prior to this change, we were using a flag value,
`reset_availability_last_uptime_downtime_val` to record the
timestamp to which the last_uptime and last_downtime should be
updated to. This was originally done so to avoid the values
being overwritten by a paxos update.

Now, instead of using an intermediate value, we are immediately
clearing the last_uptime and last_downtime values in
pending_pool_availability object. Since we are updating the values
in the pending object, we will not lost this information due to
an incoming paxos update.

Fixes: https://tracker.ceph.com/issues/71857

Signed-off-by: Shraddha Agrawal <shraddhaag@ibm.com>
(cherry picked from commit 950d911b4b28ffaffc2d63cfc15d4edabce2b533)