doc/mgr/crash.rst: remove outdated module enabling instructions
The crash module has been enabled by default since commit 8ffe44c5 in
Nautilus and is now in the always_on_modules list. However, the
documentation still contained instructions for manually enabling it.

When users followed these outdated instructions, they encountered:
```
  module 'crash' is already enabled (always-on)
```

The module cannot be disabled either. Running:
```
  ceph mgr module disable crash
```

Returns the error:
```
  Error EINVAL: module 'crash' cannot be disabled (always-on)
```
In this change, we remove the obsolete enabling instructions and clarify
that this module is always active and cannot be disabled.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>