rgw: fix aclRequired for bucket-logging
The aclRequired field in the bucket log is meant to indicate
that an acl was required to authorize the operation. This change
introduces a new field in the req_state to track whether an acl was
checked to authorize the request.

Fixes: https://tracker.ceph.com/issues/71730

Signed-off-by: N Balachandran <nithya.balachandran@ibm.com>
(cherry picked from commit 7ab070413cff8a6954142fa7855d6d258822eadf)