qa/workunits/rados: add ability to run rados workunit locally
This commit addes a `--vstart` flag that developers can pass
to run this script locally. The user must be in the `ceph/build`
directory, which is explained as a comment in the script.

The script prepares the test to be run locally by doing the following:
1. Compiling all the necessry targets
2. Setting up a test cluster
3. Prefixing `./bin/` to all executable commands to make them
   vstart-friendly
4. Tearing down the test cluster at the end.

Fixes: https://tracker.ceph.com/issues/70772

Signed-off-by: Laura Flores <lflores@ibm.com>
(cherry picked from commit 0eea3d53ee2c9dba519e49ae1e66f5493d10388f)