mgr/smb: add support for tls credential resoruces to sqlite store
The sqlite store needs to be updated for each top-level resource type.
Update it for the newly added tls credential type. Configure it so that
it works similarly to the join auth resource such that the cert data is
not stored in the sqlite db - only in the layered mon store.

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit b5e0ab8547b8faf058df24c4e5077886dec02d24)