mgr/smb: add a RawConfigEntry protocol type
Previously, the ConfigEntry type was created to be general interface for
serializing structures to JSON and persist them in a store.
However, there are times we want to retain data in a store that is not
serialized JSON - but just raw data. Create a new protocol for that
purpose.

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 6d9e126eb25625dec5a2de61955ad625a6303d1a)