mgr/smb: add a uri lookup func to the MonKeyConfigStore
Add a new lookup_uri function to the MonKeyConfigStore - this allows the
store to return an entry given a URI. The URIs in the stores are
typically used to communicate to components outside the mgr module. But
there are occasions that we have a uri and want to look it up instead
of using a key.

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 2789802fc147fb79c8ae28fecfa974db2f165721)