mgr/smb: move functions from handler.py to staging.py
Create a new file staging.py to reduce the size of handler.py and
organize this a bit more. The staging.py file will be responsible for
the special staging area store as well as functions that work directly
on the staging store like the cross-check of resources.

This change is nearly a straight more except for renaming some functions
from _foo to foo.

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 978e92fb68a1717a2713017382bfb5fbe9a24e63)