mgr/smb: add a new BigString helper type for serializing yaml
Add a new BigString type that serves to help serialize resources to
YAML, causing it to use the multi-line literal style. A BigString is
a string so it should be transparent to other serializations, but
note that it should only be used in "simplified" structures.

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 39f73e2b7926e2f717947243a24c8743e4dcb815)