mgr/smb: add a new generalized enum for source references
The join auth and users and groups sources typically specify a
source as a reference to a (local) resource. Right now this is
really the only supported value. Since they're all basically the same
add a new generic SourceReferenceType enum for future source
resource components.

Note that python doesn't allow subclassing enums that have values
assigned. This is the reason for the copy paste here.

Signed-off-by: John Mulligan <jmulligan@redhat.com>