squash! rgw: SSE-KMS Secrets Cache: use unique keys in keyring
There is a race condition between lookup_or and cache eviction that
can cause two call once functions to run for the same key.

If they were to both add that secret under the same key to the kernel
keystore the second call would only update the key. Afterwards we can
end up with multiple keystore RAII wrappers deleting the same key.

To work around this add a uuid "version" to the keyring secrets key.