install-deps.sh: use gcc-12 for ubuntu jammy
Clang relies on the system's libstdc++ headers for standard C++
features. On Ubuntu Jammy, gcc-11 is the default and it's libstdc++ is
missing some of the newer features supported by C++20.

This causes build failures as certain standard headers are incomplete.
To fix this, we install gcc-12 explicitly so that Clang can pick up the
newer libstdc++ headers.

Signed-off-by: Aishwarya Mathuria <amathuri@redhat.com>