do_cmake: Skip finding gcc if flags are already set
When compiling with clang, one would explicitly set the CMAKE COMPILER
flags. Later on when do_cmake would run we would still try to find gcc
version - and append the gcc compiler flags on top of the existing
clang falgs.
This behavior is safe since the prior flags would be used.
However, this could be confusing and probably error prone.

For example, our make check is setting the compiler to clang-19:
```
CI_DEBUG: Finding compiler for ci-build
CI_DEBUG: Our cmake_opts are:  -DCMAKE_CXX_COMPILER=clang++-19
-DCMAKE_C_COMPILER=clang-19
```
and later when running do_camke the result is:
```
+ cmake -GNinja -DWITH_PYTHON3=3.10 -DWITH_CCACHE=ON
  -DCMAKE_CXX_COMPILER=g++-11 -DCMAKE_C_COMPILER=gcc-11
  -DCMAKE_CXX_COMPILER=clang++-19 -DCMAKE_C_COMPILER=clang-19
  -- The CXX compiler identification is Clang 19.1.7
  -- The C compiler identification is Clang 19.1.7
  -- The ASM compiler identification is Clang with GNU-like command-line
  -- Found assembler: /usr/bin/clang-19
```

Signed-off-by: Matan Breizman <mbreizma@redhat.com>