osd: EC optimizations proc_master_log bug after PG merge
Boundary case bug in proc_master_log for optimized EC
pools - when searching for the first divergent entry
it didn't cope with the entire log being the divergence
and all entries being partial writes. Normally this
doesn't happen because the 1st write cannot be a partial
write, it can happen after a PG merge or after the log
is trimmed.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>