rgw: don't enable versioning on bucket when enabling snapshots
Originally we had set bucket with versioning-suspended if versioning
was not enabled previously. However, this can be problematic for
a few reasons. Setting versioning-suspended is cross-site operation so
will modify the behavior on other sites. Also, this will turn regular
object deletes into delete-marker creation, which can be problematic
as users don't really expect that.

Signed-off-by: Yehuda Sadeh <ysadehwe@ibm.com>