qa/config/crimson_qa_overrides: increase complaint time
The following is true only for tentacle:
We're unlikely to backport the fix of https://tracker.ceph.com/issues/69612
to tentacle.
The only 2 failures we see in tentacle runs are of the tracker above.
This commit doubles the complaint time in order to keep tentacle runs
fully green.
If decided to backport the fix for specific reasons, we could revert
this change.

Signed-off-by: Matan Breizman <mbreizma@redhat.com>