mgr/cephadm: unset MON flag during upgrade
This commit unsets the specified mon flags before an upgrade
starts and sets them after an upgrade.

For full upgrade, it unsets the flag before upgrade and resets
it after upgrade is complete.

For staggered upgrade, it will unset and set the flag for each
section of upgrade done separately.

If upgrade is stopped by the user, the flag will be reset before
upgrade stops.

If the upgrade fails, the flag will be reset before upgrade
exists.

Fixes: https://tracker.ceph.com/issues/72489
Signed-off-by: Shraddha Agrawal <shraddhaag@ibm.com>

Signed-off-by: Shraddha Agrawal <shraddha.agrawal000@gmail.com>