qa/standalone/scrub: fix "scrubbed in 0ms" in osd-scrub-test.sh
The specific test looks for a 'last scrub duration' higher than
0 as a sign that the scrub actually ran.  Previous code fixes
guaranteed that even a scrub duration as low as 1ms would be
reported as "1" (1s).  However, none of the 15 objects created
in this test were designated for the tested PG, which remained
empty.  As a result, the scrub duration was reported as "0".

The fix is to create a large enough number of objects so that
at least one of them is mapped to the tested PG.

Fixes: https://tracker.ceph.com/issues/71801
Signed-off-by: Ronen Friedman <rfriedma@redhat.com>
(cherry picked from commit b303afed7a8b2a65043f56170ed478f8d2bc591a)