rgw: identify case where need to return -ENOENT
When snap_id=1 the instance is in the olh. This means that we don't
try to look at the snap_map when getting the object state (because
we don't try to follow_olh. This can be a problem if the specific
instance doesn't exist (either was never created, or the snapshot was
deleted).

Signed-off-by: Yehuda Sadeh <ysadehwe@ibm.com>